Pi-side/requirements.txt
halo==0.0.31
numpy==1.20.3
tensorflow==1.14.0
opencv_python==4.5.1.48
imagezmq==1.1.1
paho_mqtt==1.5.1
RPi==0.0.1
